Remove useless function in `overview` module
The `check_update_redis_lb` was used to check that the time of
the player in Redis and MariaDB were the same before building
the records array for the `/overview` request, so that it updates
the Redis leaderboard if so.

However, even if they were, the `get_rank` function from the
`records_lib::ranks` module already checks that inconsistency.