fix: make sort order based on first or last inputs
When providing the `first` input (e.g. `first: 10`), the order is
ascending (or, for latest records, descending). If providing the `last`
input, the order is descending (or, for latest records, ascending). If
none of them is provided, it defaults to an ascending order. If both are
provided, the behavior is undefined.

Remove useless GraphQL sort type, and only provide a `sortField:
MapRecordSortableField` as input of the `map.recordsConnection` GraphQL
node.

Also add missing `EventEditionMap.recordsConnection` GraphQL node.

Refs: #97.